Make ConnectionInterface use LoggerInterface instance.
Update the internal class QueryLogger to implement LoggerInterface.
Query interpolation code has been moved to LoggedQuery itself.
